# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"AppController","description":"Rest controller on warehouse manager project"}],"paths":{"/app/profile":{"get":{"tags":["AppController"],"summary":"Getting authenticated profile usercard","operationId":"profile_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usercard"}}}}}},"put":{"tags":["AppController"],"summary":"Updating authenticated profile userecard","operationId":"profile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usercard"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usercard"}}}}}}},"/app/profile/{id}":{"get":{"tags":["AppController"],"summary":"Get usercard of user with id(in credentials table)","operationId":"profile_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usercard"}}}}}},"put":{"tags":["AppController"],"summary":"Change usercard of user with id(in credentials table)","operationId":"fillUsercard","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usercard"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usercard"}}}}}}},"/app/good":{"get":{"tags":["AppController"],"summary":"Get number of good from rack with address","operationId":"getGoodByAddr","parameters":[{"name":"addr","in":"query","required":true,"schema":{"type":"string"}},{"name":"number","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Good"}}}}}},"put":{"tags":["AppController"],"summary":"Put number of good to rack with address","operationId":"putGoodByAddr","parameters":[{"name":"addr","in":"query","required":true,"schema":{"type":"string"}},{"name":"number","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/app/good-info/{id}":{"get":{"tags":["AppController"],"summary":"Getting general info about good with id","operationId":"good","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Good"}}}}}},"put":{"tags":["AppController"],"summary":"DEPRECATED!Manual changing good in goods table","operationId":"changeGood","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Good"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Good"}}}}}},"post":{"tags":["AppController"],"summary":"DEPRECATED!Manual adding good in goods table","operationId":"newGood","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Good"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Good"}}}}}},"delete":{"tags":["AppController"],"summary":"DEPRECATED!Manual deleting good in goods table","operationId":"delGood","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/app/bind-rack":{"put":{"tags":["AppController"],"summary":"Bind rack on good with id","operationId":"bindRack","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"addr","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/app/rack":{"post":{"tags":["AppController"],"summary":"set up new rack with address","operationId":"createRack","parameters":[{"name":"addr","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["AppController"],"summary":"delete rack by address","operationId":"delRack","parameters":[{"name":"addr","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Rack"}}}}}}},"/app/racks":{"get":{"tags":["AppController"],"summary":"get info about all racks","operationId":"racks","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Rack"}}}}}}}},"/app/racks/{id}":{"get":{"tags":["AppController"],"summary":"get info about all racks binded on good with id","operationId":"goodRacks","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Rack"}}}}}}}},"/app/queue":{"get":{"tags":["AppController"],"summary":"get all queue logs","operationId":"queue","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Outcoming"}}}}}}}},"/app/profiles":{"get":{"tags":["AppController"],"summary":"Getting all profiles usercards(info), request sender has to have all roles","operationId":"profiles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Usercard"}}}}}}}},"/app/logs":{"get":{"tags":["AppController"],"summary":"get all logs","operationId":"logs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LogEntry"}}}}}}}},"/app/goods":{"get":{"tags":["AppController"],"description":"Getting all goods information","operationId":"goods","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Good"}}}}}}}},"/app/good/{id}":{"get":{"tags":["AppController"],"summary":"getting good from racks","description":"searching all rack associated with this good and sequentially get from each \ntil they run out or the requested amount is reached","operationId":"getGood","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"number","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Good"}}}}}}}},"components":{"schemas":{"Usercard":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"lastname":{"type":"string"},"firstname":{"type":"string"},"birthday":{"type":"string","format":"date-time"},"roles":{"type":"string"}}},"Good":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"descr":{"type":"string"},"totalnumber":{"type":"integer","format":"int32"}}},"Rack":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"addr":{"type":"string"},"good":{"$ref":"#/components/schemas/Good"},"number":{"type":"integer","format":"int32"}}},"Outcoming":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"usercard":{"$ref":"#/components/schemas/Usercard"},"datetime":{"type":"string","format":"date-time"},"rack":{"$ref":"#/components/schemas/Rack"},"good":{"$ref":"#/components/schemas/Good"},"number":{"type":"integer","format":"int32"},"satisfied":{"type":"boolean"}}},"LogEntry":{"type":"object"}}}}